Add test-emacs.sh
This script concatenates its arguments to a podman/docker-run
execution which mounts the package root into the container and changes
the working directory to the mount. We also attempt to cache Elpa
packages per Emacs release.

This can be changed in future to fix SELinux systems with
Podman (--security-opt=label=disable for home directory sources) and
warn users if we wish to do Docker-in-Docker testing and we mount the
dockerd socket inside the container.