Add Makefile
At the moment the Makefile contains only the rule to make the build
image, in future we can kick off tests in the Makefile from CI to
reduce the drift between local and remote development. My internet is
not so great and I hate waiting for things.

Fix makefile rule name

Rejig makefile rule names